Refactor chat engines with shared base
