+++ repo_agent/chat_engine_base.py
@@ -1,0 +1,135 @@
+from __future__ import annotations
+
+from typing import Type
+
+from llama_index.core.llms.base import LLM
+
+from repo_agent.doc_meta_info import DocItem
+from repo_agent.log import logger
+from repo_agent.prompt import chat_template
+from repo_agent.settings import SettingsManager
+
+
+class BaseChatEngine:
+    """Shared chat engine logic for different LLM backends."""
+
+    def __init__(self, llm_cls: Type[LLM], project_manager):
+        self.settings = SettingsManager.get_setting()
+        self.llm = llm_cls(
+            api_key=self.settings.chat_completion.openai_api_key.get_secret_value(),
+            api_base=self.settings.chat_completion.openai_base_url,
+            request_timeout=self.settings.chat_completion.request_timeout,
+            model=self.settings.chat_completion.model,
+            temperature=self.settings.chat_completion.temperature,
+            max_retries=1,
+            is_chat_model=True,
+        )
+
+    def build_prompt(self, doc_item: DocItem):
+        """Build the prompt messages used for the LLM call."""
+        code_info = doc_item.content
+        is_referenced = len(doc_item.who_reference_me) > 0
+
+        code_type = code_info["type"]
+        code_name = code_info["name"]
+        code_content = code_info["code_content"]
+        have_return = code_info.get("have_return", False)
+        file_path = doc_item.get_full_name()
+
+        def get_referenced_prompt(doc_item: DocItem) -> str:
+            if len(doc_item.reference_who) == 0:
+                return ""
+            prompt = [
+                """As you can see, the code calls the following objects, their code and docs are as following:"""
+            ]
+            for reference_item in doc_item.reference_who:
+                doc_text = reference_item.md_content[-1] if reference_item.md_content else "None"
+                code_snippet = reference_item.content.get("code_content", "")
+                instance_prompt = (
+                    f"obj: {reference_item.get_full_name()}\n"
+                    f"Document: \n{doc_text}\n"
+                    f"Raw code:```\n{code_snippet}\n```" + "=" * 10
+                )
+                prompt.append(instance_prompt)
+            return "\n".join(prompt)
+
+        def get_referencer_prompt(doc_item: DocItem) -> str:
+            if len(doc_item.who_reference_me) == 0:
+                return ""
+            prompt = [
+                """Also, the code has been called by the following objects, their code and docs are as following:"""
+            ]
+            for referencer_item in doc_item.who_reference_me:
+                doc_text = referencer_item.md_content[-1] if referencer_item.md_content else "None"
+                code_snippet = referencer_item.content.get("code_content", "None")
+                instance_prompt = (
+                    f"obj: {referencer_item.get_full_name()}\n"
+                    f"Document: \n{doc_text}\n"
+                    f"Raw code:```\n{code_snippet}\n```" + "=" * 10
+                )
+                prompt.append(instance_prompt)
+            return "\n".join(prompt)
+
+        def get_relationship_description(referencer_content, reference_letter):
+            if referencer_content and reference_letter:
+                return "And please include the reference relationship with its callers and callees in the project from a functional perspective"
+            elif referencer_content:
+                return "And please include the relationship with its callers in the project from a functional perspective."
+            elif reference_letter:
+                return "And please include the relationship with its callees in the project from a functional perspective."
+            else:
+                return ""
+
+        code_type_tell = "Class" if code_type == "ClassDef" else "Function"
+        parameters_or_attribute = "attributes" if code_type == "ClassDef" else "parameters"
+        have_return_tell = (
+            "**Output Example**: Mock up a possible appearance of the code's return value."
+            if have_return
+            else ""
+        )
+        combine_ref_situation = (
+            "and combine it with its calling situation in the project,"
+            if is_referenced
+            else ""
+        )
+
+        referencer_content = get_referencer_prompt(doc_item)
+        reference_letter = get_referenced_prompt(doc_item)
+        has_relationship = get_relationship_description(referencer_content, reference_letter)
+
+        project_structure_prefix = ", and the related hierarchical structure of this project is as follows (The current object is marked with an *):"
+
+        return chat_template.format_messages(
+            combine_ref_situation=combine_ref_situation,
+            file_path=file_path,
+            project_structure_prefix=project_structure_prefix,
+            code_type_tell=code_type_tell,
+            code_name=code_name,
+            code_content=code_content,
+            have_return_tell=have_return_tell,
+            has_relationship=has_relationship,
+            reference_letter=reference_letter,
+            referencer_content=referencer_content,
+            parameters_or_attribute=parameters_or_attribute,
+            language=self.settings.project.language,
+        )
+
+    def generate_doc(self, doc_item: DocItem):
+        """Generate documentation for the provided DocItem."""
+        messages = self.build_prompt(doc_item)
+        try:
+            logger.debug(f"Sending messages to LLM: {messages}")
+            response = self.llm.chat(messages)
+
+            if hasattr(response.raw, "usage") and isinstance(response.raw.usage, dict):
+                logger.debug(f"LLM Prompt Tokens: {response.raw.usage.get('prompt_tokens')}")
+                logger.debug(f"LLM Completion Tokens: {response.raw.usage.get('completion_tokens')}")
+                logger.debug(f"Total LLM Token Count: {response.raw.usage.get('total_tokens')}")
+            else:
+                logger.debug("No usage information available in response.")
+
+            logger.debug(f"LLM Response: {response.message.content}")
+            return response.message.content
+        except Exception as e:
+            logger.error(f"Error in llamaindex chat call: {e}")
+            raise
